last seen  11/10/2024

+++ lib/features/food_feature/restaurant_dashboard/presentation/pages/restaurant_dashboard_view.dart
@@ -3,15 +3,14 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:fourtyninehub/common/widgets/stateless/buttons/app_button.dart';
-import 'package:fourtyninehub/common/widgets/stateless/dynamic/shared_scaffold.dart';
 import 'package:fourtyninehub/common/widgets/stateless/labels/label.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:fourtyninehub/features/food_feature/create_restaurant/cubit/create_resturant_cubit.dart';
 import 'package:fourtyninehub/features/food_feature/create_restaurant/views/create_resturant_view.dart';
 import 'package:fourtyninehub/features/food_feature/restaurant_dashboard/presentation/widgets/restaurant_statistics_view.dart';
+import 'package:fourtyninehub/features/food_feature/restaurants_list/data/models/is_restaurant_model.dart';
+import 'package:fourtyninehub/features/food_feature/restaurants_list/presentation/cubit/restaurants_list_cubit.dart';
 import 'package:fourtyninehub/res/style/app_colors.dart';
-import 'package:fourtyninehub/routes/routes.dart';
-import 'package:go_router/go_router.dart';
 
 import '../../../../../common/widgets/dynamic/sizer.dart';
 import '../../../../../res/strings/labels.dart';
@@ -21,18 +20,14 @@ import '../cubit/restaurant_dashboard_cubit.dart';
 import '../widgets/restaurant_order_card.dart';
 
 class RestaurantDashboardView extends StatelessWidget {
-  const RestaurantDashboardView({super.key});
+  final IsRestaurantModel isRestaurantModel;
+
+  const RestaurantDashboardView({super.key, required this.isRestaurantModel});
 
   @override
   Widget build(BuildContext context) {
-    final String? restaurantId =
-    ModalRoute
-        .of(context)
-        ?.settings
-        .arguments as String?;
-
-    final controller = context.read<RestaurantDashboardCubit>()
-      ..loadData();
+    print("${isRestaurantModel.restaurantId}asfdfadsfdsfas");
+    final controller = context.read<RestaurantDashboardCubit>()..loadData();
     return BlocConsumer<RestaurantDashboardCubit, RestaurantDashboardState>(
         listener: (context, state) {},
         builder: (context, state) {
@@ -51,28 +46,47 @@ class RestaurantDashboardView extends StatelessWidget {
                   children: [
                     Padding(
                       padding: const EdgeInsets.symmetric(horizontal: 4.0),
-                      child: Row(
-                        children: [
-                          Expanded(
-                              child: Label(
-                                text: state.connected
+                      child:
+                          BlocConsumer<RestaurantsCubit, RestaurantsListState>(
+                        listener: (context, state) {
+                          // TODO: implement listener
+                        },
+                        builder: (context, state) {
+                          return Row(
+                            children: [
+                              Expanded(
+                                  child: Label(
+                                text: context
+                                            .watch<RestaurantsCubit>()
+                                            .state
+                                            .isResturant!
+                                            .isActive ??
+                                        false
                                     ? Labels.connected
                                     : Labels.notConnected,
                                 style: Styles.headerText(),
                               )),
-                          if (state.connected)
-                            SizedBox(
-                              height: 15.h,
-                              width: 15.w,
-                              child: const CircularProgressIndicator.adaptive(),
-                            ),
-                          Switch(
-                              value: state.connected,
-                              inactiveThumbColor: Colors.white,
-                              inactiveTrackColor: Colors.grey,
-                              onChanged: (v) async =>
-                              await controller.changeConnectivityStatus())
-                        ],
+                              // if (state.connected)
+                              //   SizedBox(
+                              //     height: 15.h,
+                              //     width: 15.w,
+                              //     child: const CircularProgressIndicator.adaptive(),
+                              //   ),
+                              Switch(
+                                  value: context
+                                          .watch<RestaurantsCubit>()
+                                          .state
+                                          .isResturant!
+                                          .isActive ??
+                                      isRestaurantModel.isActive!,
+                                  inactiveThumbColor: Colors.white,
+                                  inactiveTrackColor: Colors.grey,
+                                  onChanged: (v) async => await context
+                                      .read<RestaurantsCubit>()
+                                      .changeConnectivityStatus(v))
+                            ],
+                          );
+                        },
                       ),
                     ),
                     // Divider(),
@@ -96,12 +110,12 @@ class RestaurantDashboardView extends StatelessWidget {
                                     MaterialPageRoute(
                                       builder: (context) =>
                                           BlocProvider<CreateRestaurantCubit>(
-                                            create: (context) =>
-                                                serviceLocator(),
-                                            child: const CreateRestaurantForm(
-                                              from: 'update',
-                                            ),
-                                          ),
+                                        create: (context) => serviceLocator(),
+                                        child: CreateRestaurantForm(
+                                            from: 'update',
+                                            restaurantId:
+                                                isRestaurantModel.restaurantId),
+                                      ),
                                     ));
                                 // context.push(Routes.CREATERESTURANT);
                               },
@@ -115,8 +129,19 @@ class RestaurantDashboardView extends StatelessWidget {
                             child: AppButton(
                               label: 'Delete Registration',
                               onPressed: () {
-                                controller.deleteRestaurantById(
-                                    id: restaurantId!);
+                                _showConfirmationDialog(
+                                  context,
+                                  title: "Delete Restaurant",
+                                  message:
+                                      "Are you sure you want to remove this restaurant?",
+                                  onConfirm: () async {
+                                    // Perform the logout action here
+                                    await controller.deleteRestaurantById(
+                                        context,
+                                        id: isRestaurantModel.restaurantId!);
+                                    Navigator.pop(context);
+                                  },
+                                );
                               },
                               backColor: AppColors.PRIMARY_COLOR_DARK,
                               style: Styles.headerText(color: Colors.white),
@@ -129,28 +154,27 @@ class RestaurantDashboardView extends StatelessWidget {
                     Expanded(
                       child: ListView.separated(
                           itemBuilder: (context, index) {
-                            print(state.orders!.data.orders.length.toString() +
-                                '455555555555555555555555');
+                            print(
+                                '${state.orders!.data.orders.length}455555555555555555555555');
                             return Column(
                               children: [
                                 RestaurantOrderCard(
                                     item: state.orders!.data.orders[index]),
                                 state.orders!.data.restaurantSubscriptionType ==
-                                    'Not subscribed'
+                                        'Not subscribed'
                                     ? Text(
-                                  'Please Subscribe to contact the client!',
-                                  style: Styles.headerText(
-                                      color:
-                                      AppColors.PRIMARY_COLOR_DARK),
-                                )
+                                        'Please Subscribe to contact the client!',
+                                        style: Styles.headerText(
+                                            color:
+                                                AppColors.PRIMARY_COLOR_DARK),
+                                      )
                                     : const Sizer(),
                               ],
                             );
                           },
-                          separatorBuilder: (context, index) =>
-                          const Sizer(
-                            height: 20,
-                          ),
+                          separatorBuilder: (context, index) => const Sizer(
+                                height: 20,
+                              ),
                           itemCount: state.orders?.data.orders.length ?? 0),
                     ),
                   ],
@@ -158,4 +182,44 @@ class RestaurantDashboardView extends StatelessWidget {
               ));
         });
   }
+}
+
+// Generalized function to show the confirmation dialog
+void _showConfirmationDialog(
+  BuildContext context, {
+  required String title,
+  required String message,
+  required VoidCallback onConfirm,
+  VoidCallback? onCancel,
+}) {
+  showDialog(
+    context: context,
+    builder: (BuildContext context) {
+      return AlertDialog(
+        title: Text(title),
+        content: Text(message),
+        actions: <Widget>[
+          TextButton(
+            onPressed: onCancel ?? () => Navigator.of(context).pop(),
+            child: Text("No"),
+          ),
+          ElevatedButton(
+            onPressed: onConfirm,
+            child: Text("Yes"),
+          ),
+        ],
+      );
+    },
+  );
+}
+
+// Example function to handle the deletion
+void _deleteItem(BuildContext context) {
+  // Close the dialog
+  Navigator.of(context).pop();
+
+  // Show a snackbar after deletion
+  ScaffoldMessenger.of(context).showSnackBar(
+    SnackBar(content: Text("Item deleted successfully")),
+  );
 }
